+++ src/administrators/views.py
@@ -63,8 +63,10 @@ class AdministratorModelViewSet(ModelViewSet, EventLogMixin):
             administrator.last_name = last_name
             administrator.save()
 
-            self.log_event(request, operated_object=user)
-            self.log_event(request, operated_object=administrator)
+            validated_data = serializer.validated_data
+
+            self.log_event(request, user, validated_data=validated_data)
+            self.log_event(request, administrator, validated_data=validated_data)
 
             return Response(
                 AdministratorSerializer(administrator).data,
@@ -80,9 +82,11 @@ class AdministratorModelViewSet(ModelViewSet, EventLogMixin):
         serializer = self.get_serializer(
             instance, data=request.data, partial=partial
         )
+
         if serializer.is_valid():
+            validated_data = serializer.validated_data
             self.perform_update(serializer)
-            self.log_event(request, operated_object=instance)
+            self.log_event(request, instance, validated_data=validated_data)
 
             return Response(serializer.data, status=status.HTTP_200_OK)
         else:
@@ -108,7 +112,7 @@ class AdministratorModelViewSet(ModelViewSet, EventLogMixin):
             instance.status = ProfileStatus.DEACTIVATED
             instance.save()
 
-            self.log_event(request, operated_object=instance)
+            self.log_event(request, instance)
 
             return Response(
                 {"detail": "Object deactivated successfully."},